Refactored to use Grid instead of size and nMines (dependency injection)

+++ src/main/java/it/unibo/pps/e3/Logics.java
@@ -2,7 +2,7 @@ package it.unibo.pps.e3;
 
 public interface Logics {
 
-    int getNumberOfCells();
+    long getNumberOfCells();
 
-    int countMines();
+    long countMines();
 }
